Only download data that does not yet exist.

+++ utils/dataio.py
@@ -1,6 +1,7 @@
 import requests
 import json
 import pkgutil
+import os
 
 def get_data(day, year = 2018):
     url = get_url(day, year)
@@ -17,11 +18,11 @@ def get_url(day, year = 2018):
 
 def load_data(day, year = 2018):
     filename = ''.join(['data//day', str(day), '.dat'])
-    # if the file does not exist
-    data = get_data(day, year)
-    with open( filename , 'wt') as file:
-        file.write(data)
-    # else
-    with open( filename , 'r') as file:
-        data = file.read()
-    return data
+    if not os.path.exists(filename):
+        data = get_data(day, year)
+        with open( filename , 'wt') as file:
+            file.write(data)
+    else:
+        with open( filename , 'r') as file:
+            data = file.read()
+        return data
